+++ src/main/java/com/projet/scootop/scootop/controller/ApiServiceInteraction.java
@@ -3,8 +3,6 @@ package com.projet.scootop.scootop.controller;
 import com.projet.scootop.scootop.model.services.interaction.AlertDTO;
 import com.projet.scootop.scootop.model.services.interaction.FollowDTO;
 import com.projet.scootop.scootop.model.services.interaction.InviteDTO;
-import com.projet.scootop.scootop.service.domainuser.ChairmanService;
-import com.projet.scootop.scootop.service.domainuser.PlayerService;
 import com.projet.scootop.scootop.service.services.interaction.AlertService;
 import com.projet.scootop.scootop.service.services.interaction.FollowService;
 import com.projet.scootop.scootop.service.services.interaction.InviteService;
@@ -27,7 +25,7 @@ public class ApiServiceInteraction {
 
 
     @GetMapping("api/alert/{id}")
-    AlertDTO getAlert(@PathVariable int id){
+    AlertDTO getAlert(@PathVariable Long id){
         return alertService.get(id);
     }
     @PostMapping("api/alert")
@@ -35,11 +33,11 @@ public class ApiServiceInteraction {
         return alertService.add(body);
     }
     @PutMapping("api/alert/{id}")
-    String updateAlert(@RequestBody String body,@PathVariable int id){
+    String updateAlert(@RequestBody String body,@PathVariable Long id){
         return body;
     }
     @DeleteMapping("api/alert/{id}")
-    String deleteAlert(@PathVariable int id){
+    String deleteAlert(@PathVariable Long id){
         return alertService.delete(id);
     }
     @GetMapping("api/alerts")
@@ -48,7 +46,7 @@ public class ApiServiceInteraction {
     }
 
     @GetMapping("api/follow/{id}")
-    FollowDTO getFollow(@PathVariable int id){
+    FollowDTO getFollow(@PathVariable Long id){
         return followService.get(id);
     }
     @PostMapping("api/follow")
@@ -56,11 +54,11 @@ public class ApiServiceInteraction {
         return followService.add(body);
     }
     @PutMapping("api/follow/{id}")
-    String updateFollow(@RequestBody String body,@PathVariable int id){
+    String updateFollow(@RequestBody String body,@PathVariable Long id){
         return body;
     }
     @DeleteMapping("api/follow/{id}")
-    String deleteFollow(@PathVariable int id){
+    String deleteFollow(@PathVariable Long id){
         return followService.delete(id);
     }
     @GetMapping("api/follows")
@@ -69,7 +67,7 @@ public class ApiServiceInteraction {
     }
 
     @GetMapping("api/invite/{id}")
-    InviteDTO getInvite(@PathVariable int id){
+    InviteDTO getInvite(@PathVariable Long id){
         return inviteService.get(id);
     }
     @PostMapping("api/invite")
@@ -77,11 +75,11 @@ public class ApiServiceInteraction {
         return inviteService.add(body);
     }
     @PutMapping("api/invite/{id}")
-    String updateInvite(@RequestBody String body,@PathVariable int id){
+    String updateInvite(@RequestBody String body,@PathVariable Long id){
         return body;
     }
     @DeleteMapping("api/invite/{id}")
-    String deleteInvite(@PathVariable int id){
+    String deleteInvite(@PathVariable Long id){
         return inviteService.delete(id);
     }
     @GetMapping("api/invites")
